# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Connect API
  version: 1.0.0
  description: |
    Production REST API specification for **Social Connect**, an Instagram-inspired full-stack social media application built with the MERN stack.
    
    ### Key Architecture & Security Features:
    - **Dual Token Authentication**: Short-lived JWT Access Tokens (`Authorization: Bearer <token>`) combined with rotated, HttpOnly Refresh Tokens (`refreshToken` cookie).
    - **Session Invalidation & Replay Detection**: Refresh tokens are family-tracked with automatic revocation of all sessions upon reuse attack detection.
    - **Bilateral Interaction Protections**: Immediate blocking enforcement across feeds, user search, direct messaging, comments, and stories.
    - **Content Moderation State Machine**: Multi-stage reporting queue and non-destructive moderation statuses (`ACTIVE`, `HIDDEN`, `REMOVED`).
    - **Deterministic Cursor-Based Pagination**: Stable chronological feed pagination utilizing compound `(createdAt, _id)` database indexes without offset drift.
    - **Immutable Audit Logging**: Compliance-grade, non-repudiable audit logs tracking authentication, administrative interventions, and security anomalies.
servers:
  - url: http://localhost:5000/api
    description: Local Development Server
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [Authentication & Sessions]
      summary: Register new user account
      description: Creates a new user profile with hashed password and dispatches an email verification token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '409':
          $ref: '#/components/responses/409Conflict'
  /auth/login:
    post:
      operationId: loginUser
      tags: [Authentication & Sessions]
      summary: Authenticate user credentials
      description: Verifies email/password, issues a short-lived Access Token in response body, and sets a secure HttpOnly Refresh Token cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=abc123xyz...; Path=/api/auth; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          description: Account suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [Authentication & Sessions]
      summary: Rotate and refresh access token
      description: Presents the HttpOnly refresh token cookie (or x-refresh-token header). Issues a new access token and rotates the refresh token. Detects replay attacks and revokes the session family if reused.
      parameters:
        - in: header
          name: x-refresh-token
          required: false
          schema:
            type: string
          description: Alternative header if client cannot set cookies
      responses:
        '200':
          description: Token rotated and refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          $ref: '#/components/responses/401Unauthorized'
...
  /auth/sessions:
    get:
      operationId: getSessions
      tags: [Authentication & Sessions]
      summary: List active login sessions
      security:
        - bearerAuth: []
      description: Returns list of active device sessions for the authenticated user with userAgent and IP metadata.
      responses:
        '200':
          description: Active sessions returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
  /auth/sessions/{sessionId}:
    delete:
      operationId: revokeSession
      tags: [Authentication & Sessions]
      summary: Revoke specific login session
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '404':
          $ref: '#/components/responses/404NotFound'
  /auth/change-password:
    patch:
      operationId: changePassword
      tags: [Authentication & Sessions]
      summary: Change password
      security:
        - bearerAuth: []
      description: Updates password for authenticated user and revokes all existing sessions except the current one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          $ref: '#/components/responses/400BadRequest'
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [Authentication & Sessions]
      summary: Request password reset email
      description: Generates a cryptographically secure token and sends a password reset link to user's registered email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Reset email dispatched (or generic success to avoid enumeration)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [Authentication & Sessions]
      summary: Complete password reset
      description: Verifies reset token hash, updates password, and revokes all active user sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          $ref: '#/components/responses/400BadRequest'
  /auth/verify-email:
    post:
      operationId: verifyEmail
      tags: [Authentication & Sessions]
      summary: Verify email address
      description: Consumes the email verification token and marks account emailVerified=true.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  example: f8d9c2e0b1a2...
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          $ref: '#/components/responses/400BadRequest'
  /auth/resend-verification:
    post:
      operationId: resendVerification
      tags: [Authentication & Sessions]
      summary: Resend email verification link
      security:
        - bearerAuth: []
      description: Dispatches a new email verification token if the authenticated user's email is unverified.
      responses:
        '200':
          description: Verification email resent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
...
  /users/me/blocked:
    get:
      operationId: getBlockedUsers
      tags: [Blocks]
      summary: List blocked users
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageQueryParam'
        - $ref: '#/components/parameters/LimitQueryParam'
      responses:
        '200':
          description: List of blocked users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
...
  /users/{username}/block:
    post:
      operationId: blockUser
      tags: [Blocks]
      summary: Block a user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernamePathParam'
      responses:
        '200':
          description: User blocked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '400':
          description: Cannot block self
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: User already blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      operationId: unblockUser
      tags: [Blocks]
      summary: Unblock a user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernamePathParam'
      responses:
        '200':
          description: User unblocked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
  /users/{username}/block-status:
    get:
      operationId: getBlockStatus
      tags: [Blocks]
      summary: Check block status with a user
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UsernamePathParam'
      responses:
        '200':
          description: Block relationship status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Standard JWT access token passed in Authorization header
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: Secure HttpOnly refresh token cookie
    skipRateLimit:
      type: apiKey
      in: header
      name: x-skip-rate-limit
      description: Bypasses rate limiting in testing environments
  parameters:
    PageQueryParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-indexed page number
    LimitQueryParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
      description: Items per page
    UsernamePathParam:
      name: username
      in: path
      required: true
      schema:
        type: string
        minLength: 3
        maxLength: 30
      description: Target username (case-insensitive)
...
    SessionIdParam:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      description: 24-character hexadecimal MongoDB ObjectId of the session
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
        data:
          type: object
          nullable: true
    ErrorEnvelope:
      type: object
      required: [success, message, error]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Resource not found
        error:
          type: object
          properties:
            code:
              type: string
              example: NOT_FOUND
            message:
              type: string
              example: Resource not found
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          example: alex_chen
        email:
          type: string
          format: email
          example: alex@example.com
        password:
          type: string
          minLength: 8
          maxLength: 128
          example: SecretPass123!
        fullName:
          type: string
          maxLength: 60
          example: Alex Chen
        bio:
          type: string
          maxLength: 150
          example: Full-stack developer and photographer
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: alex@example.com
        password:
          type: string
          example: SecretPass123!
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          example: OldPassword123!
        newPassword:
          type: string
          minLength: 8
          maxLength: 128
          example: NewSecretPass123!
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
          example: f9d8a7b6c5...
        newPassword:
          type: string
          minLength: 8
          maxLength: 128
          example: FreshSecretPass123!
...
  responses:
    400BadRequest:
      description: Invalid request body, malformed query, or validation failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Validation failed
            error:
              code: VALIDATION_ERROR
              message: Invalid input fields
    401Unauthorized:
      description: Authentication missing, invalid token, or revoked session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Authentication required
            error:
              code: UNAUTHORIZED
              message: Missing or invalid token
...
    404NotFound:
      description: Target resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: Resource not found
            error:
              code: NOT_FOUND
              message: The requested resource does not exist
    409Conflict:
      description: Resource conflict or unique constraint violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            message: A user with this email or username already exists
            error:
              code: CONFLICT
              message: Duplicate resource